installer & cegui/tinyxml

+++ Recipes/cegui.py
@@ -8,20 +8,21 @@ import util
 
 hg_repo = 'http://crayzedsgui.hg.sourceforge.net/hgroot/crayzedsgui/cegui_mk2'
 
-@simple.source('cegui', "Crazy Eddie's GUI")
+@simple.source('cegui', "CEGUI")
 def recipe(package):
     path = package[standard.a_location]
-    #if exists(path):
-    #    print 'Removing old files...'
-    #    util.rmtree(path)
+    if exists(path):
+        print 'Removing old files...'
+        util.rmtree(path)
 
-    #util.run('hg clone %s "%s"' % (hg_repo, path))
-
+    util.run('hg clone %s "%s"' % (hg_repo, path))
+    
     build = join(path, 'build')
-    if exists(build):
-        print 'Removing build dir...'
-        util.rmtree(build)
-    os.makedirs(build)
+    if 0:
+        if exists(build):
+            print 'Removing build dir...'
+            util.rmtree(build)
+        os.makedirs(build)
 
     options = [
         ('CEGUI_BUILD_XMLPARSER_TINYXML', '1'),
@@ -32,10 +33,10 @@ def recipe(package):
     options = ' '.join('-D %s=%s' % (k, v) for k, v in options)
 
     util.run('cmake -G"NMake Makefiles" %s ..' % options, cwd=build)
+    util.run('nmake', cwd=build)
 
-    raise Exception
-
-
-@environment.headers.implement(recipe)
-def find_headers(package):
-    return [join(package[standard.a_location], 'cegui', 'include')]
+environment.define_paths(recipe, {
+    environment.headers: ['cegui\\include'],
+    environment.libraries: ['build\\lib'],
+    environment.executables: ['build\\bin']
+})
